backend/requirements.txt
absl-py==0.7.1
args==0.1.0
atomicwrites==1.3.0
async-timeout==2.0.1
attrdict==2.0.0
attrs==18.2.0
backcall==0.1.0
certifi==2019.3.9
backports.functools-lru-cache==1.4
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
better-exceptions==0.2.2
bintrees==2.0.7
bitarray==0.8.3
cached-property==1.3.1
cachetools==3.1.1
chardet==3.0.4
Click==7.0
cloudpickle==1.0.0
ConfigArgParse==0.14.0
cycler==0.10.0
dask==1.2.2
dash==0.28.5
dash-core-components==0.36.1
dash-html-components==0.10.1
dash-renderer==0.14.3
dateparser==0.7.0
decorator==4.4.0
distributed==1.25.1
dlib==19.17.0
dominate==2.4.0
entrypoints==0.2.3
enum34==1.1.6
eth-abi==1.2.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-typing==1.3.0
eth-utils==1.2.1
Events==0.3
Flask==1.0.2
Flask-Caching==1.4.0
Flask-Compress==1.4.0
Flask-Cors==3.0.7
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.1
flask-marshmallow==0.9.0
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
flatbuffers==1.10
funcsigs==1.0.2
future==0.17.1
glob2==0.6
greenlet==0.4.15
grequests==0.3.0
grpcio==1.18.0
gunicorn==20.0.4
h5py==2.9.0
heapdict==1.0.0
hexbytes==0.1.0
httplib2==0.13.0
idna==2.8
imageio==2.5.0
imagesize==1.1.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
isort==4.3.4
itsdangerous==1.1.0
jsonpickle==0.9.4
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keras-rl==0.4.2
keyring==17.0.0
matplotlib==3.0.3
matlab_wrapper==1
mock==3.0.5
lazy-object-proxy==1.3.1
libarchive-c==2.8
line-profiler==2.1.2
llvmlite==0.26.0
matlab-wrapper==1
more-itertools==4.3.0
msgpack==0.5.6
msgpack-numpy==0.4.4.3
neat==0.4.1
neat-python==0.92
networkx==2.3
parso==0.5.0
pbr==5.2.0
pexpect==4.7.0
nose==1.3.7
notebook==5.7.4
numba==0.41.0
numexpr==2.6.8
opencv-python==4.1.0.25
numpy==1.16.4
packaging==18.0
pandas==0.23.4
pandas-datareader==0.7.0
pep8==1.7.1
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.11.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==5.4.8
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyglet==1.4.0b1
pymongo==3.7.2
PyMySQL==0.8.1
pyOpenSSL==18.0.0
pyparsing==2.4.0
pytest==4.4.2
PySocks==1.6.8
python-dateutil==2.8.0
python-editor==1.0.3
python-gflags==3.1.2
python-rapidjson==0.7.0
python-snappy==0.5.1
python-utils==2.3.0
pytz==2018.7
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.1
redis==2.10.6
regex==2018.2.8
requests==2.21.0
requests-file==1.4.2
requests-ftp==0.3.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
scikit-image==0.15.0
scikit-learn==0.21.0
scipy==1.2.1
seaborn==0.9.0
service-identity==17.0.0
simplegeneric==0.8.1
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.12.0
sortedcollections==1.0.1
sortedcontainers==2.1.0
SQLAlchemy==1.2.15
sqlparse==0.3.0
statsmodels==0.9.0
tables==3.4.4
tabulate==0.8.2
termcolor==1.1.0
toolz==0.9.0
tornado==5.1.1
tox==2.9.1
tqdm
traitlets==4.3.2
translate==3.5.0
translate-api==2.1.0
Twisted==17.9.0
typing==3.6.4
typing-extensions==3.6.6
uritemplate==3.0.0
urllib3==1.24.3
virtualenv==15.2.0
virtualenv-clone==0.5.1
wcwidth==0.1.7
Werkzeug==0.15.2
webencodings==0.5.1
WebOb==1.8.4
websocket-client==0.40.0
websockets==6.0
ws4py==0.5.1
